fix conversion functions to use appropriate vtk function

+++ modules/viz/include/opencv2/viz/viz3d.hpp
@@ -44,8 +44,8 @@ namespace cv
             Affine3f getViewerPose();
             void setViewerPose(const Affine3f &pose);
             
-            void convertToWindowCoordinates(const Point3f &pt, Point3f &window_coord);
-            void converTo3DRay(const Point3f &window_coord, Point3f &origin, Vec3f &direction);
+            void convertToWindowCoordinates(const Point3d &pt, Point3d &window_coord);
+            void converTo3DRay(const Point3d &window_coord, Point3d &origin, Vec3d &direction);
 
             void spin();
             void spinOnce(int time = 1, bool force_redraw = false);
