+++ modules/viz/src/viz3d_impl.hpp
@@ -126,8 +126,8 @@ public:
     void setViewerPose(const Affine3f &pose);
     Affine3f getViewerPose();
 
-    void convertToWindowCoordinates(const Point3f &pt, Point3f &window_coord);
-    void converTo3DRay(const Point3f &window_coord, Point3f &origin, Vec3f &direction);
+    void convertToWindowCoordinates(const Point3d &pt, Point3d &window_coord);
+    void converTo3DRay(const Point3d &window_coord, Point3d &origin, Vec3d &direction);
